requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
bcrypt==4.0.1
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
databases==0.8.0
ecdsa==0.18.0
elastic-transport==8.12.0
elasticsearch==8.12.0
fastapi==0.109.2
fastapi-jwt-auth==0.5.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
multidict==6.0.5
orjson==3.9.14
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==1.7.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.51
SQLAlchemy-Utils==0.41.1
starlette==0.36.3
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
pandas==2.2.1
openpyxl==3.1.2
email-validator==2.1.1
